Car Rental System
The basic functions of an online car rental system are to keep tracks of vehicles, staff, customers and booking. It provides useful information to the staff such as giving daily reports of vehicles to be delivered/picked up and acts as a vehicle management system by monitoring the use and price of the vehicles.